+++ lex-web-ui/test/unit/index.js
@@ -9,5 +9,10 @@ testsContext.keys().forEach(testsContext);
 // require all src files except main.js for coverage.
 // you can also change this to match only the subset of files that
 // you want coverage for.
-const srcContext = require.context('../../src', true, /^\.\/(?!main(\.js)?$)/);
+const srcContext = require.context(
+  '../../src',
+  true,
+  // lex-web-ui: added LexApp.vue to exception
+  /^\.\/(?!main(\.js)?|(LexApp(\.vue)?)$)/,
+);
 srcContext.keys().forEach(srcContext);